requirements.txt
alembic==1.7.4
aniso8601==9.0.1
attrs==21.2.0
cffi==1.15.0
click==8.0.3
coverage==6.1.1
cryptography==35.0.0
Flask==2.0.2
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.2
Mako==1.1.5
MarkupSafe==2.0.1
marshmallow==3.14.0
marshmallow-sqlalchemy==0.26.1
packaging==21.2
pluggy==1.0.0
py==1.10.0
pycparser==2.20
PyJWT==2.3.0
PyMySQL==1.0.2
pyparsing==2.4.7
pytest==6.2.5
python-dotenv==0.19.1
pytz==2021.3
simplejson==3.17.5
six==1.16.0
SQLAlchemy==1.4.26
toml==0.10.2
watchdog==2.1.6
Werkzeug==2.0.2
